backend/pyproject.toml
[project]
name = "booktracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.5",
    "djangorestframework>=3.15.2",
    "pillow>=11.1.0",
]

[dependency-groups]
dev = [
    "django-types>=0.20.0",
]
